broke out App into its own component
